added force_refresh option to -html plugin
html plugin was not pulling the most recent version of the html document
if the document was already in the browser cache. if force_refresh is
set to true, then the current timestamp is appended on as a URL argument
as a means to force the browser to grab the newest version.