call-func fix; added active-match
call-function: changed data storage so that if there was no return value
then no data would be stored in the data object. this helps when trying
to write data to a database as the empty data can be safely ignored.

active-match: new plugin for manipulating an image until it matches
another image.