added options to si-kr plugin
can now opt to leave the stimulus on the screen for a fixed length of
time, still recording the responses when they occur.

also allows a timelimit to be set on the responses, so that if no
responses is recorded after some period of time, the trial completes and
records a -1 as the response.