Fix #532 to allow script on external-html pages to be executed optionally in a external-html plugin trial
Issue #532 addresses the issue that scripts included in the remote page will not be executed via the XMLHttpRequest() in the external-html script.

We will need to relocate the content of all scripts by relocating the content through DOM manipulation. The execution is optional and can be controlled through the new execute_Script parameter.