Accuracy information in lexical decision experiment
Hello,

Congrats and thank you for this great library.

Please consider this bit of code for the lexical decision experiment. It adds a bit of information about accuracy to the results feedback--for instance, 'Your accuracy rate was 95% (38 correct trials, 2 incorrect, and 0 unanswered)'. 

No worries at all if you'd rather not apply it, or if you'd like to make any edits.

Thanks